move tabs to constant directory
