+++ src/constants/tabs.ts
@@ -1,10 +1,10 @@
-import EditIcon from "../../../public/assets/icons/settings_icon.svg"
-import NewIcon from "../../../public/assets/icons/folder_icon.svg"
-import StatsIcon from "../../../public/assets/icons/chart_icon.svg"
-import HomeIcon from "../../../public/assets/icons/home_icon.svg"
-import DashboardIcon from "../../../public/assets/icons/event_icon.svg"
-import ProfileIcon from "../../../public/assets/icons/profile_icon.svg"
-import ExitIcon from "../../../public/assets/icons/exit_icon.svg"
+import EditIcon from "../../public/assets/icons/settings_icon.svg"
+import NewIcon from "../../public/assets/icons/folder_icon.svg"
+import StatsIcon from "../../public/assets/icons/chart_icon.svg"
+import HomeIcon from "../../public/assets/icons/home_icon.svg"
+import DashboardIcon from "../../public/assets/icons/event_icon.svg"
+import ProfileIcon from "../../public/assets/icons/profile_icon.svg"
+import ExitIcon from "../../public/assets/icons/exit_icon.svg"
 
 export const admin = [
     {
